+++ src/app/(sst)/school-of-technology/layout.tsx
@@ -1,4 +1,3 @@
-import { Suspense } from 'react';
 import type { Metadata } from 'next';
 
 import { AnnouncementStrip, Header } from '@components/common';
@@ -8,15 +7,6 @@ import { AlumniProvider } from '@modules/sst/alumni-directory/context/AlumniCont
 import { METADATA } from '@utils/common/metadata';
 import { ANNOUNCEMENT_STRIP_CONTENT } from '@utils/sst/constants';
 
-import {
-  Analytics,
-  AnalyticsFallback,
-  MicrosoftClarity,
-} from '@/components/common/Analytics';
-import {
-  PRODUCTS,
-  SUB_PRODUCTS,
-} from '@/components/common/Analytics/constants';
 
 export const metadata: Metadata = METADATA.SST;
 
@@ -32,16 +22,7 @@ export default function Layout({ children }: { children: React.ReactNode }) {
         />
         <Navbar />
       </Header>
-
       <main>{children}</main>
-
-      <Suspense key="gtm-script" fallback={<AnalyticsFallback />}>
-        <Analytics
-          product={PRODUCTS.SCHOOL_OF_TECHNOLOGY}
-          subProduct={SUB_PRODUCTS.ALUMNI_DIRECTORY}
-        />
-        <MicrosoftClarity />
-      </Suspense>
     </AlumniProvider>
   );
 }
